+++ js/options.js
@@ -1,0 +1,1 @@
+export { defaultOptions, contextlessOptions } from './fp/options.js'